feat: API CLIENT for axios calling, refactor code structure
