+++ src/app/(auth)/_layout.tsx
@@ -1,0 +1,12 @@
+import {Stack} from "expo-router";
+
+const AuthLayout = () => {
+  return (
+    <Stack>
+      <Stack.Screen name="login" options={{headerShown: false}}/>
+      <Stack.Screen name="register" options={{headerShown: false}}/>
+    </Stack>
+  )
+}
+
+export default AuthLayout;
