+++ src/app/(tabs)/index.tsx
@@ -1,7 +1,6 @@
 import {StyleSheet} from "react-native";
 import {useTheme} from "@/theme";
 import SafeAreaView from '@/components/global/SafeAreaView'
-import Text from "@/components/global/Text";
 
 const Map = () => {
   const {colors} = useTheme()
@@ -9,9 +8,6 @@ const Map = () => {
 
   return (
     <SafeAreaView>
-      <Text>
-        Open up app/index.tsx to start working app!
-      </Text>
     </SafeAreaView>
   )
 }
