+++ src/app/_layout.tsx
@@ -23,10 +23,7 @@ const RootLayoutNav = () => {
         <Stack.Screen name="users/[userId]" options={{headerShown: false}}/>
       </Stack.Protected>
       <Stack.Protected guard={!isAuthenticated}>
-        <Stack.Screen name="login" options={{headerShown: false}}/>
-      </Stack.Protected>
-      <Stack.Protected guard={!isAuthenticated}>
-        <Stack.Screen name="register" options={{headerShown: false}}/>
+        <Stack.Screen name="(auth)" options={{headerShown: false}}/>
       </Stack.Protected>
     </Stack>
   );
