--- src/app/login.tsx
@@ -1,227 +1,0 @@
-import React, {useState} from 'react';
-import {
-  ActivityIndicator,
-  Alert,
-  KeyboardAvoidingView,
-  Platform,
-  ScrollView,
-  StyleSheet,
-  TouchableOpacity,
-  View,
-} from 'react-native';
-import {Link} from 'expo-router';
-import {useAuth} from '@/contexts/AuthContext';
-import {useTheme} from '@/theme';
-import {validateEmail} from '@/utils/auth';
-import {MaterialIcons} from "@expo/vector-icons/";
-import {SafeAreaView, Text, TextInput} from "@/components/global";
-
-const Login = () => {
-  const {login, isLoading} = useAuth();
-  const {colors, isDark} = useTheme();
-
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [errors, setErrors] = useState<{ email?: string; password?: string }>({});
-  const [isSecurePassword, setIsSecurePassword] = useState(true);
-
-  const handleLogin = async () => {
-    setErrors({});
-    const newErrors: { email?: string; password?: string } = {};
-    const trimmedPassword = password.trim();
-
-    if (!email) {
-      newErrors.email = 'Email is required';
-    } else if (!validateEmail(email)) {
-      newErrors.email = 'Please enter a valid email';
-    }
-
-    if (!trimmedPassword) {
-      newErrors.password = 'Password is required';
-    }
-
-    if (Object.keys(newErrors).length > 0) {
-      setErrors(newErrors);
-      return;
-    }
-
-    try {
-      const response = await login(email.trim().toLowerCase(), trimmedPassword);
-      if (response.status === 401) setErrors({
-        email: 'Invalid email or password',
-        password: 'Invalid email or password'
-      });
-    } catch (error: any) {
-      Alert.alert('Login Failed', error.message);
-    }
-  };
-
-  const styles = StyleSheet.create({
-    scrollContent: {
-      flexGrow: 1,
-      justifyContent: 'center',
-      padding: 24,
-    },
-    header: {
-      marginBottom: 48,
-      alignItems: 'center',
-    },
-    title: {
-      fontSize: 36,
-      fontWeight: 'bold',
-      color: colors.text,
-      marginBottom: 8,
-      fontFamily: 'Whyte',
-    },
-    subtitle: {
-      fontSize: 16,
-      color: colors.text,
-      opacity: 0.7,
-    },
-    form: {
-      marginBottom: 24,
-    },
-    inputContainer: {
-      marginBottom: 20,
-    },
-    label: {
-      fontSize: 14,
-      fontWeight: '600',
-      color: colors.text,
-      marginBottom: 8,
-      opacity: 0.8,
-    },
-    input: {
-      backgroundColor: isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.05)',
-      borderRadius: 12,
-      padding: 16,
-      fontSize: 16,
-      color: colors.text,
-      borderWidth: 1,
-      borderColor: isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.1)',
-    },
-    inputError: {
-      borderColor: '#ff4444',
-    },
-    errorText: {
-      color: '#ff4444',
-      fontSize: 12,
-      marginTop: 4,
-    },
-    button: {
-      backgroundColor: colors.primary,
-      borderRadius: 12,
-      padding: 16,
-      alignItems: 'center',
-      marginTop: 8,
-    },
-    buttonDisabled: {
-      opacity: 0.6,
-    },
-    buttonText: {
-      color: '#ffffff',
-      fontSize: 16,
-      fontWeight: 'bold',
-    },
-    footer: {
-      alignItems: 'center',
-      marginTop: 24,
-    },
-    footerText: {
-      fontSize: 14,
-      color: colors.text,
-      opacity: 0.7,
-    },
-    link: {
-      color: colors.primary,
-      fontWeight: 'bold',
-    },
-    securePassword: {
-      position: 'absolute',
-      top: 42.5,
-      right: 17.5,
-    }
-  });
-
-  return (
-
-    <SafeAreaView>
-      <KeyboardAvoidingView
-        behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
-        style={{flex: 1}}
-      >
-        <ScrollView
-          contentContainerStyle={styles.scrollContent}
-          keyboardShouldPersistTaps="handled"
-        >
-          <View style={styles.header}>
-            <MaterialIcons name='share-location' size={128} color={colors.primary}/>
-          </View>
-
-          <View style={styles.form}>
-            <View style={styles.inputContainer}>
-              <Text style={styles.label}>Email</Text>
-              <TextInput
-                style={[styles.input, errors.email && styles.inputError]}
-                placeholder="Enter your email"
-                value={email}
-                onChangeText={setEmail}
-                autoCapitalize="none"
-                keyboardType="email-address"
-                editable={!isLoading}
-                accessibilityLanguage={'en-US'}
-                maxLength={50}
-              />
-              {errors.email && <Text style={styles.errorText}>{errors.email}</Text>}
-            </View>
-
-            <View style={styles.inputContainer}>
-              <Text style={styles.label}>Password</Text>
-              <TextInput
-                style={[styles.input, errors.password && styles.inputError]}
-                placeholder="Enter your password"
-                placeholderTextColor={isDark ? 'rgba(255, 255, 255, 0.4)' : 'rgba(0, 0, 0, 0.4)'}
-                value={password}
-                onChangeText={setPassword}
-                secureTextEntry={isSecurePassword}
-                editable={!isLoading}
-                accessibilityLanguage={'en-US'}
-                autoCapitalize='none'
-                maxLength={50}
-              />
-              <MaterialIcons name={isSecurePassword ? 'visibility' : 'visibility-off'} size={24} color={colors.text}
-                             onPress={() => setIsSecurePassword(!isSecurePassword)}
-                             style={styles.securePassword}
-              />
-              {errors.password && <Text style={styles.errorText}>{errors.password}</Text>}
-            </View>
-
-            <TouchableOpacity
-              style={[styles.button, isLoading && styles.buttonDisabled]}
-              onPress={handleLogin}
-              disabled={isLoading}
-            >
-              {isLoading ? (
-                <ActivityIndicator color="#ffffff"/>
-              ) : (
-                <Text style={styles.buttonText}>Sign In</Text>
-              )}
-            </TouchableOpacity>
-          </View>
-
-          <View style={styles.footer}>
-            <Text style={styles.footerText}>
-              Don't have an account?{' '}
-              <Link href="/register" style={styles.link}>
-                Sign Up
-              </Link>
-            </Text>
-          </View>
-        </ScrollView>
-      </KeyboardAvoidingView>
-    </SafeAreaView>
-
-  );
-};
-
-export default Login;
