+++ src/components/MapOld/utils/movementStateUtils.ts
