+++ src/components/MapOld/utils/validators.ts
@@ -1,5 +1,3 @@
-import { STATE_STABILITY_CONFIG } from "./constants";
-
 export const validateStateChange = async (
   currentState: string,
   newState: string,
